userinfo recoil 적용
